Change references.js location to util folder
