+++ src/components/ADempiere/DataTable/menu/menuTableMixin.js
@@ -1,7 +1,7 @@
 import { supportedTypes, exportFileFromJson, exportFileZip } from '@/utils/ADempiere/exportUtil'
 import { showNotification } from '@/utils/ADempiere/notification'
 import { recursiveTreeSearch } from '@/utils/ADempiere/valueUtils'
-import { FIELDS_QUANTITY } from '@/components/ADempiere/Field/references'
+import { FIELDS_QUANTITY } from '@/utils/ADempiere/references'
 
 export const menuTableMixin = {
   props: {
